Fixes for irrationals in all Julia versions
`zero(x)`, `one(x)`, and `inv(x)` were not implemented for `x`
irrational in old Julia version. `oneunit(x)` is not implemented for `x`
irrational in any Julia version (as of Julia 1.11). `impl_$f(x)` are
introduced for all these functions to fix the compatibility and provide
suitable equivalent to `inv(x)`.